toml_datetime::de

Enum VisitMap

pub enum VisitMap<'de> {
    Datetime(Datetime),
    Key(Cow<'de, str>),
}

Integrate Datetime into an untagged deserialize

Variants

Datetime(Datetime)

The map was deserialized as a Datetime value

Key(Cow<'de, str>)

The map is of an unknown format and needs further deserialization

Implementations

impl<'de> VisitMap<'de>

pub fn next_key_seed<V: MapAccess<'de>>( visitor: &mut V, ) -> Result<Option<Self>, V::Error>

Determine the type of the map by deserializing it