Struct toml_datetime::Time

pub struct Time {
    pub hour: u8,
    pub minute: u8,
    pub second: u8,
    pub nanosecond: u32,
}

A parsed TOML time value

May be part of a Datetime. Alone, Time corresponds to a Local Time. From the TOML v1.0.0 spec:

If you include only the time portion of an RFC 3339 formatted date-time, it will represent that time of day without any relation to a specific day or any offset or timezone.

lt1 = 07:32:00
lt2 = 00:32:00.999999

Millisecond precision is required. Further precision of fractional seconds is implementation-specific. If the value contains greater precision than the implementation can support, the additional precision must be truncated, not rounded.

Fields
hour: u8

Hour: 0 to 23

minute: u8

Minute: 0 to 59

second: u8

Second: 0 to {58, 59, 60} (based on leap second rules)

nanosecond: u32

Nanosecond: 0 to 999_999_999